Update SAiOSAdPlugin.m
bannerViewDidLoadAd method not firing iAdBannerViewDidLoadAdEventHandler js function, so changed it to be similar to the working didFailToReceiveAdWithError method below.

I suck with objective C and never contributed anything to a project before so I apologise if this is badly implemented but that small change was enough for me to get the important iAdBannerViewDidLoadAdEventHandler function working again. Hopefully that might help someone.